--- js/random-number.js
@@ -1,8 +1,0 @@
-const getArbitraryNumber = (min, max, decimalPlaces) => {
-  if (max <= min || max <= 0 || min < 0) {
-    return 'Некорректный ввод'
-  }
-  return (Math.random() * (max - min) + min).toFixed(decimalPlaces);
-};
-
-export {getArbitraryNumber};
