+++ js/random-numbers.js
@@ -1,3 +1,10 @@
+const getArbitraryNumber = (min, max, decimalPlaces) => {
+  if (max <= min || max <= 0 || min < 0) {
+    return 'Некорректный ввод'
+  }
+  return (Math.random() * (max - min) + min).toFixed(decimalPlaces);
+};
+
 const getInteger = (min, max) => {
   if (max <= min || max <= 0 || min < 0) {
     return 'Некорректный ввод'
@@ -7,4 +14,5 @@ const getInteger = (min, max) => {
   return Math.floor(Math.random() * (max - min + 1)) + min;
 };
 
+export {getArbitraryNumber};
 export {getInteger};
