+++ js/total-array.js
@@ -1,11 +1,11 @@
-import {getObject} from './object.js';  
+import {getObjectAd} from './object.js';  
 
 const NUMBER_OF_OBJECTS = 10;
 
 const getArray = () => {
   let array = [];
   for(let i = 0; i < NUMBER_OF_OBJECTS; i++) {
-    array.push(getObject());
+    array.push(getObjectAd());
   }    return array;
 };
